Implement new Observation granular permissions.
Issue: IBP-5924/IBP-6238
Review: none

+++ src/main/java/org/ibp/api/rest/dataset/DatasetResource.java
@@ -14,6 +14,7 @@ import org.generationcp.middleware.domain.etl.MeasurementVariable;
 import org.generationcp.middleware.domain.inventory.common.SearchCompositeDto;
 import org.generationcp.middleware.domain.ontology.VariableType;
 import org.generationcp.middleware.manager.api.SearchRequestService;
+import org.generationcp.middleware.pojos.workbench.PermissionsEnum;
 import org.generationcp.middleware.service.api.dataset.FilteredPhenotypesInstancesCountDTO;
 import org.generationcp.middleware.service.api.dataset.ObservationUnitEntryReplaceRequest;
 import org.generationcp.middleware.service.api.dataset.ObservationUnitsParamDTO;
@@ -38,6 +39,7 @@ import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -48,6 +50,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import springfox.documentation.annotations.ApiIgnore;
 
+import javax.servlet.http.HttpServletRequest;
 import java.io.File;
 import java.util.Arrays;
 import java.util.List;
@@ -88,6 +91,9 @@ public class DatasetResource {
 
 	@Autowired
 	private DatasetLock datasetLock;
+
+	@Autowired
+	private HttpServletRequest request;
 
 	@ApiOperation(value = "Get Dataset Columns", notes = "Retrieves ALL MeasurementVariables (columns) associated to the dataset, "
 		+ "that will be shown in the Observation Table")
@@ -182,9 +188,7 @@ public class DatasetResource {
 	}
 
 	@ApiOperation(value = "Add Observation", notes = "Add Observation")
-	@PreAuthorize("hasAnyAuthority('ADMIN','STUDIES', 'MANAGE_STUDIES','CREATE_STUDIES', 'DELETE_STUDY', 'CLOSE_STUDY', 'LOCK_STUDY','MS_MANAGE_OBSERVATION_UNITS','MS_WITHDRAW_INVENTORY','MS_CREATE_PENDING_WITHDRAWALS', "
-		+ "'MS_CREATE_CONFIRMED_WITHDRAWALS', 'MS_CANCEL_PENDING_TRANSACTIONS', 'MS_MANAGE_FILES','MS_CREATE_LOTS', 'MS_GERMPLASM_AND_CHECKS', 'MS_VIEW_GERMPLASM_AND_CHECKS', 'MS_ADD_ENTRY_DETAILS_VARIABLES', 'MS_MODIFY_ENTRY_DETAILS_VALUES', "
-		+ "'MS_MODIFY_COLUMNS', 'MS_REPLACE_GERMPLASM', 'MS_ADD_NEW_ENTRIES', 'MS_IMPORT_ENTRY_DETAILS')")
+	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES', 'MS_OBSERVATIONS', 'MS_MANAGE_CONFIRMED_OBSERVATIONS')")
 	@RequestMapping(value = "/{crop}/programs/{programUUID}/studies/{studyId}/datasets/{datasetId}/observationUnits/{observationUnitId}/observations", method = RequestMethod.POST)
 	public ResponseEntity<ObservationDto> addObservation(
 		@PathVariable final String crop, @PathVariable final String programUUID, @PathVariable final Integer studyId,
@@ -196,14 +200,15 @@ public class DatasetResource {
 	}
 
 	@ApiOperation(value = "Update Observation", notes = "Update Observation")
-	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES', 'CREATE_STUDIES', 'DELETE_STUDY', 'CLOSE_STUDY', 'LOCK_STUDY', 'MS_MANAGE_OBSERVATION_UNITS', 'MS_WITHDRAW_INVENTORY', 'MS_CREATE_PENDING_WITHDRAWALS', "
-		+ "'MS_CREATE_CONFIRMED_WITHDRAWALS', 'MS_CANCEL_PENDING_TRANSACTIONS', 'MS_MANAGE_FILES', 'MS_CREATE_LOTS', 'MS_GERMPLASM_AND_CHECKS', 'MS_VIEW_GERMPLASM_AND_CHECKS', 'MS_ADD_ENTRY_DETAILS_VARIABLES', 'MS_MODIFY_ENTRY_DETAILS_VALUES', "
-		+ "'MS_MODIFY_COLUMNS', 'MS_REPLACE_GERMPLASM', 'MS_ADD_NEW_ENTRIES', 'MS_IMPORT_ENTRY_DETAILS')")
 	@RequestMapping(value = "/{crop}/programs/{programUUID}/studies/{studyId}/datasets/{datasetId}/observationUnits/{observationUnitId}/observations/{observationId}", method = RequestMethod.PATCH)
 	public ResponseEntity<ObservationDto> updateObservation(
 		@PathVariable final String crop, @PathVariable final String programUUID, @PathVariable final Integer studyId,
 		@PathVariable final Integer datasetId, @PathVariable final Integer observationUnitId, @PathVariable final Integer observationId,
 		@ApiParam("Only some fields will be updated: ie. value, draftValue") @RequestBody final ObservationDto observationDto) {
+
+		if (!this.hasAuthority(observationDto.isDraftMode())) {
+			throw new AccessDeniedException("");
+		}
 
 		return new ResponseEntity<>(
 			this.studyDatasetService.updateObservation(studyId, datasetId, observationId, observationUnitId, observationDto),
@@ -225,7 +230,7 @@ public class DatasetResource {
 	}
 
 	@ApiOperation(value = "Generate and save a sub-observation dataset", notes = "Returns the basic information for the generated dataset")
-	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES')")
+	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES','MS_STUDY_ACTIONS','MS_MANAGE_OBSERVATION_UNITS','MS_CREATE_SUB_OBSERVATION_UNITS')")
 	@RequestMapping(value = "/{cropName}/programs/{programUUID}/studies/{studyId}/datasets/{parentId}/generation", method = RequestMethod.POST)
 	@ResponseBody
 	public ResponseEntity<DatasetDTO> generateDataset(@PathVariable final String cropName, @PathVariable final String programUUID,
@@ -342,7 +347,7 @@ public class DatasetResource {
 	}
 
 	@ApiOperation(value = "Delete Observation", notes = "Delete Observation")
-	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES')")
+	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES', 'MS_OBSERVATIONS', 'MS_MANAGE_CONFIRMED_OBSERVATIONS')")
 	@RequestMapping(
 		value = "/{crop}/programs/{programUUID}/studies/{studyId}/datasets/{datasetId}/observationUnits/{observationUnitId}/observations/{observationId}",
 		method = RequestMethod.DELETE)
@@ -421,7 +426,7 @@ public class DatasetResource {
 	}
 
 	@ApiOperation(value = "Move draft value to saved value in sub-observation dataset", notes = "Save information for the imported dataset")
-	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES')")
+	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES', 'MS_OBSERVATIONS', 'MS_ACCEPT_PENDING_OBSERVATION')")
 	@RequestMapping(value = "/{crop}/programs/{programUUID}/studies/{studyId}/datasets/{datasetId}/observation-units/drafts/acceptance", method = RequestMethod.POST)
 	@ResponseBody
 	public ResponseEntity<Void> acceptDraftData(@PathVariable final String crop, @PathVariable final String programUUID,
@@ -433,7 +438,7 @@ public class DatasetResource {
 	}
 
 	@ApiOperation(value = "Reject draft value in sub-observation dataset", notes = "Reject information for the imported dataset")
-	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES')")
+	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES', 'MS_OBSERVATIONS', 'MS_ACCEPT_PENDING_OBSERVATION')")
 	@RequestMapping(value = "/{crop}/programs/{programUUID}/studies/{studyId}/datasets/{datasetId}/observation-units/drafts/rejection", method = RequestMethod.POST)
 	@ResponseBody
 	public ResponseEntity<Void> rejectDraftData(@PathVariable final String crop, @PathVariable final String programUUID,
@@ -460,7 +465,7 @@ public class DatasetResource {
 	}
 
 	@ApiOperation(value = "Set missing value to saved value in sub-observation dataset", notes = "Set missing for the imported dataset")
-	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES')")
+	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES', 'MS_OBSERVATIONS', 'MS_ACCEPT_PENDING_OBSERVATION')")
 	@RequestMapping(value = "/{crop}/programs/{programUUID}/studies/{studyId}/datasets/{datasetId}/observation-units/drafts/set-as-missing", method = RequestMethod.POST)
 	@ResponseBody
 	public ResponseEntity<Void> setValuesToMissing(@PathVariable final String crop, @PathVariable final String programUUID,
@@ -485,7 +490,7 @@ public class DatasetResource {
 	}
 
 	@ApiOperation(value = "Move draft value to saved value in sub-observation dataset", notes = "Save information for the imported dataset")
-	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES')")
+	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES', 'MS_OBSERVATIONS', 'MS_ACCEPT_PENDING_OBSERVATION')")
 	@RequestMapping(value = "/{crop}/programs/{programUUID}/studies/{studyId}/datasets/{datasetId}/observation-units/drafts/filter/acceptance", method = RequestMethod.POST)
 	@ResponseBody
 	public ResponseEntity<Void> acceptDraftDataByVariable(
@@ -498,7 +503,6 @@ public class DatasetResource {
 	}
 
 	@ApiOperation(value = "Set value to the selected variable", notes = "Set value to the selected variable")
-	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES')")
 	@RequestMapping(value = "/{crop}/programs/{programUUID}/studies/{studyId}/datasets/{datasetId}/observation-units/filter/set-value", method = RequestMethod.POST)
 	@ResponseBody
 	public ResponseEntity<Void> setValueToVariable(
@@ -506,12 +510,14 @@ public class DatasetResource {
 		@PathVariable final Integer datasetId,
 		@RequestBody final ObservationUnitsParamDTO paramDTO) {
 
+		if (!this.hasAuthority(paramDTO.getObservationUnitsSearchDTO().getDraftMode())) {
+			throw new AccessDeniedException("");
+		}
 		this.studyDatasetService.setValueToVariable(studyId, datasetId, paramDTO);
 		return new ResponseEntity<>(HttpStatus.OK);
 	}
 
 	@ApiOperation(value = "Delete values of the selected variable", notes = "Delete values of the selected variable")
-	@PreAuthorize("hasAnyAuthority('ADMIN', 'STUDIES', 'MANAGE_STUDIES')")
 	@RequestMapping(value = "/{crop}/programs/{programUUID}/studies/{studyId}/datasets/{datasetId}/observation-units/filter/delete-value", method = RequestMethod.POST)
 	@ResponseBody
 	public ResponseEntity<Void> deleteVariableValues(
@@ -519,6 +525,9 @@ public class DatasetResource {
 		@PathVariable final Integer datasetId,
 		@RequestBody final ObservationUnitsSearchDTO observationUnitsSearchDTO) {
 
+		if (!this.hasAuthority(observationUnitsSearchDTO.getDraftMode())) {
+			throw new AccessDeniedException("");
+		}
 		this.studyDatasetService.deleteVariableValues(studyId, datasetId, observationUnitsSearchDTO);
 		return new ResponseEntity<>(HttpStatus.OK);
 	}
@@ -599,6 +608,24 @@ public class DatasetResource {
 
 	public ResourceBundleMessageSource getMessageSource() {
 		return this.messageSource;
+	}
+
+	private boolean hasAuthority(final boolean draftMode) {
+		if (draftMode) {
+			return this.request.isUserInRole(PermissionsEnum.ADMIN.name())
+				|| this.request.isUserInRole(PermissionsEnum.STUDIES.name())
+				|| this.request.isUserInRole(PermissionsEnum.MANAGE_STUDIES.name())
+				|| this.request.isUserInRole(PermissionsEnum.MS_OBSERVATIONS.name())
+				|| this.request.isUserInRole(PermissionsEnum.MS_MANAGE_PENDING_OBSERVATIONS.name());
+		} else {
+			return this.request.isUserInRole(PermissionsEnum.ADMIN.name())
+				|| this.request.isUserInRole(PermissionsEnum.STUDIES.name())
+				|| this.request.isUserInRole(PermissionsEnum.MANAGE_STUDIES.name())
+				|| this.request.isUserInRole(PermissionsEnum.MS_OBSERVATIONS.name())
+				|| this.request.isUserInRole(PermissionsEnum.MS_MANAGE_CONFIRMED_OBSERVATIONS.name())
+				|| this.request.isUserInRole(PermissionsEnum.MS_ACCEPT_PENDING_OBSERVATION.name());
+		}
+
 	}
 
 }
